pages/setup/bluetooth: Display errors for scanning
